Added test class for pit api and not use other class
